+++ src/main/java/com/plzgraduate/myongjigraduatebe/auth/api/signin/SignInApiPresentation.java
@@ -1,9 +1,10 @@
-package com.plzgraduate.myongjigraduatebe.auth.adaptor.in.web.signin;
+package com.plzgraduate.myongjigraduatebe.auth.api.signin;
 
 import javax.validation.Valid;
 
 import org.springframework.web.bind.annotation.RequestBody;
 
+import com.plzgraduate.myongjigraduatebe.auth.api.signin.dto.request.SignInRequest;
 import com.plzgraduate.myongjigraduatebe.auth.application.port.in.TokenResponse;
 import com.plzgraduate.myongjigraduatebe.core.meta.LoginUser;
