+++ src/main/java/com/plzgraduate/myongjigraduatebe/auth/api/signin/SignInController.java
@@ -1,4 +1,4 @@
-package com.plzgraduate.myongjigraduatebe.auth.adaptor.in.web.signin;
+package com.plzgraduate.myongjigraduatebe.auth.api.signin;
 
 import javax.validation.Valid;
 
@@ -7,6 +7,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 
+import com.plzgraduate.myongjigraduatebe.auth.api.signin.dto.request.SignInRequest;
 import com.plzgraduate.myongjigraduatebe.auth.application.port.in.signin.SignInUseCase;
 import com.plzgraduate.myongjigraduatebe.auth.application.port.in.TokenResponse;
 import com.plzgraduate.myongjigraduatebe.core.meta.LoginUser;
