+++ src/main/java/com/plzgraduate/myongjigraduatebe/auth/api/token/TokenController.java
@@ -1,4 +1,4 @@
-package com.plzgraduate.myongjigraduatebe.auth.adaptor.in.web.token;
+package com.plzgraduate.myongjigraduatebe.auth.api.token;
 
 import javax.validation.Valid;
 
@@ -6,6 +6,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 
+import com.plzgraduate.myongjigraduatebe.auth.api.token.dto.request.TokenRequest;
 import com.plzgraduate.myongjigraduatebe.auth.application.port.in.AccessTokenResponse;
 import com.plzgraduate.myongjigraduatebe.auth.application.port.in.token.TokenUseCase;
 import com.plzgraduate.myongjigraduatebe.core.meta.WebAdapter;
