+++ src/test/java/com/plzgraduate/myongjigraduatebe/support/WebAdaptorTestSupport.java
@@ -13,8 +13,8 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import org.springframework.web.context.WebApplicationContext;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.plzgraduate.myongjigraduatebe.auth.adaptor.in.web.signin.SignInController;
-import com.plzgraduate.myongjigraduatebe.auth.adaptor.in.web.token.TokenController;
+import com.plzgraduate.myongjigraduatebe.auth.api.signin.SignInController;
+import com.plzgraduate.myongjigraduatebe.auth.api.token.TokenController;
 import com.plzgraduate.myongjigraduatebe.auth.application.port.in.signin.SignInUseCase;
 import com.plzgraduate.myongjigraduatebe.auth.application.port.in.token.TokenUseCase;
 import com.plzgraduate.myongjigraduatebe.auth.security.TokenProvider;
